rename and move bcd elements to match naming schema
