+++ components/compat-table/feature-row.js
