+++ components/compat-table/global.css
