+++ components/compat-table/legend.js
