+++ components/compat-table/utils.js
