--- components/compat/index.js
@@ -1,1 +1,0 @@
-export { LazyCompatTable } from "./lazy-compat-table.js";
