+++ components/content/index.js
@@ -63,7 +63,10 @@ function BCD({ id, title, query, isH3 }) {
   const level = isH3 ? 3 : 2;
   return html`<section aria-labelledby=${ifDefined(id ?? undefined)}>
     ${Heading(level, id ? String(id) : null, String(title))}
-    <lazy-compat-table locale="en-US" query=${query}></lazy-compat-table>
+    <mdn-lazy-compat-table
+      locale="en-US"
+      query=${query}
+    ></mdn-lazy-compat-table>
   </section>`;
 }
