+++ entry.client.js
@@ -3,7 +3,6 @@ import "@lit-labs/ssr-client/lit-element-hydrate-support.js";
 
 // elements:
 // TODO: migrate these all the the proper naming schema so they can be loaded by load-elements.js below
-import "./components/compat/index.js";
 import "./observatory/landing/form.js";
 import "./observatory/results/results.js";
