+++ types/element-map.d.ts
@@ -5,8 +5,10 @@ declare global {
     "mdn-button": import("../components/button/element.js").MDNButton;
     "mdn-code-example": import("../components/code-example/element.js").MDNCodeExample;
     "mdn-color-theme": import("../components/color-theme/element.js").MDNColorTheme;
+    "mdn-compat-table": import("../components/compat-table/element.js").MDNCompatTable;
     "mdn-content-feedback": import("../components/content-feedback/element.js").MDNContentFeedback;
     "mdn-copy-button": import("../components/copy-button/element.js").MDNCopyButton;
+    "mdn-lazy-compat-table": import("../components/lazy-compat-table/element.js").MDNLazyCompatTable;
     "mdn-search-modal": import("../components/search-modal/element.js").MDNSearchModal;
     "mdn-site-search": import("../components/site-search/element.js").MDNSiteSearch;
   }
